requirements.txt
# Core dependencies
aiohttp>=3.9.0
duckduckgo-search>=4.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2023.12.30
openai-whisper>=20231117
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0
requests>=2.31.0
wikipedia>=1.4.0
praw>=7.7.0
tenacity>=8.2.0
spacy>=3.7.0
datasets>=2.14.0
accelerate>=0.24.0
numpy>=1.24.0
pandas>=2.0.0

# Web application dependencies
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
plotly>=5.17.0
altair>=5.1.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0